Fragment workspace file into command chunks
